feat: allow dual module setup
