+++ src/index.test.js
@@ -2,7 +2,7 @@ import { expect, test, describe } from "bun:test";
 import { transform } from "@babel/core";
 import pluginSyntaxJsx from "@babel/plugin-syntax-jsx";
 import pluginTransReactJsx from "@babel/plugin-transform-react-jsx";
-import { autoObserverPlugin } from '../src/index.js'
+import { autoObserverPlugin } from './index.mjs'
 
 const runTransform = (input) => {
   return transform(input, {
